requirements.txt
python-telegram-bot==13.7
flask==2.0.1
requests==2.26.0
gunicorn==20.1.0
